Internal: Replace `DocumentHelper` with 'document/elements' component utils [ED-4138] (#15591)
