+++ assets/dev/js/editor/document/helper-bc.js
@@ -16,4 +16,34 @@ export default class BackwardsCompatibility {
 
 		return $e.components.get( 'document' ).utils.findContainerById( id );
 	}
+
+	static isValidChild( childModel, parentModel ) {
+		elementorCommon.helpers.softDeprecated( 'isValidChild', '3.4.0', "$e.components.get( 'document/elements' ).utils.isValidChild( childModel, parentModel )" );
+
+		return $e.components.get( 'document/elements' ).utils.isValidChild( childModel, parentModel );
+	}
+
+	static isValidGrandChild( childModel, targetContainer ) {
+		elementorCommon.helpers.softDeprecated( 'isValidGrandChild', '3.4.0', "$e.components.get( 'document/elements' ).utils.isValidGrandChild( childModel, targetContainer )" );
+
+		return $e.components.get( 'document/elements' ).utils.isValidGrandChild( childModel, targetContainer );
+	}
+
+	static isSameElement( sourceModel, targetContainer ) {
+		elementorCommon.helpers.softDeprecated( 'isSameElement', '3.4.0', "$e.components.get( 'document/elements' ).utils.isSameElement( sourceModel, targetContainer )" );
+
+		return $e.components.get( 'document/elements' ).utils.isSameElement( sourceModel, targetContainer );
+	}
+
+	static getPasteOptions( sourceModel, targetContainer ) {
+		elementorCommon.helpers.softDeprecated( 'getPasteOptions', '3.4.0', "$e.components.get( 'document/elements' ).utils.getPasteOptions( sourceModel, targetContainer )" );
+
+		return $e.components.get( 'document/elements' ).utils.getPasteOptions( sourceModel, targetContainer );
+	}
+
+	static isPasteEnabled( targetContainer ) {
+		elementorCommon.helpers.softDeprecated( 'isPasteEnabled', '3.4.0', "$e.components.get( 'document/elements' ).utils.isPasteEnabled( targetContainer )" );
+
+		return $e.components.get( 'document/elements' ).utils.isPasteEnabled( targetContainer );
+	}
 }
