+++ assets/dev/js/editor/document/hooks/data/document/elements/create/is-valid-child.js
@@ -1,5 +1,4 @@
 import Dependency from 'elementor-api/modules/hooks/data/dependency';
-import DocumentHelper from 'elementor-document/helper';
 
 export class IsValidChild extends Dependency {
 	getCommand() {
@@ -15,7 +14,7 @@ export class IsValidChild extends Dependency {
 			modelToCreate = new Backbone.Model( model );
 
 		return containers.some( ( /* Container */ container ) => (
-			DocumentHelper.isValidChild( modelToCreate, container.model )
+			$e.components.get( 'document/elements' ).utils.isValidChild( modelToCreate, container.model )
 		) );
 	}
 }
