+++ assets/dev/js/editor/document/ui/commands/paste.js
@@ -1,5 +1,4 @@
 import CommandBase from 'elementor-api/modules/command-base';
-import DocumentHelper from 'elementor-document/helper';
 
 export class Paste extends CommandBase {
 	initialize( args ) {
@@ -30,7 +29,7 @@ export class Paste extends CommandBase {
 
 		this.target.forEach( ( /* Container */ container ) => {
 			const { options = {} } = args,
-				pasteOptions = DocumentHelper.getPasteOptions( this.storage[ 0 ], container );
+				pasteOptions = $e.components.get( 'document/elements' ).utils.getPasteOptions( this.storage[ 0 ], container );
 
 			if ( ! pasteOptions.isValidChild ) {
 				if ( pasteOptions.isSameElement ) {
