+++ assets/dev/js/editor/elements/views/column-empty.js
@@ -1,5 +1,3 @@
-import DocumentHelper from 'elementor-document/helper';
-
 module.exports = Marionette.ItemView.extend( {
 	template: '#tmpl-elementor-empty-preview',
 
@@ -26,7 +24,7 @@ module.exports = Marionette.ItemView.extend( {
 					{
 						name: 'paste',
 						title: __( 'Paste', 'elementor' ),
-						isEnabled: () => DocumentHelper.isPasteEnabled( this._parent.getContainer() ),
+						isEnabled: () => $e.components.get( 'document/elements' ).utils.isPasteEnabled( this._parent.getContainer() ),
 						callback: () => $e.run( 'document/ui/paste', {
 							container: this._parent.getContainer(),
 						} ),
